Switch to Core Graphics API for event -> Unicode translation
So... I'm dumb. I was doing all the work required by UCKeyTranslate to
turn a virtual keycode into the Unicode output pressing the key should
produce, but it turns out there's already a function in Core Graphics to
do the same thing with none of the work. I saw
CGEventKeyboardGetUnicodeString when scrolling through
CGEventGetTimestamp's related functions and almost broke down as the
realization that I wasted a day implementing what should have taken five
minutes to call as a function... oh well, you know what they say. It's
not the destination that counts but the friends you made along the way.
I'll miss you TISCopyCurrentKeyboardLayoutInputSource()!