Add option to record virtual keycodes
Instead of recording the result of the keypress, this option allows for
the recording of physical keys pressed.

This breaks the v1 serialization format. However, considering the early
nature of this project, I am not guaranteeing the serialization format
will not change within a version. Once the project is released, a
serialization format will be stable.