Lazily instantiate output file
The file keystrokes are recorded in is named with a combination of the
vendor ID and product ID. The vendor ID is a value assigned by the USB
committee to the manufacturer, and the product ID is a value assigned by
the manufacturer to the device. My assumption is that these two values
together should uniquely identify the device.